fix cdn

+++ script/fetchzipai.py
@@ -11,7 +11,7 @@ HEADERS = {
 
 
 
-def article(url):
+def article(url, confirm=False):
     aid = url.split("/")[-1].split(".")[0]
     r = requests.get(ROOT+url, headers=HEADERS)
     if r.status_code != 200:
@@ -26,8 +26,11 @@ def article(url):
         if isinstance(ele, Tag):
             if ele.name == "img":
                 src = ele.get("src")
-                if not src.startswith("http"):
+                if "d/file/selfies" not in src:
                     continue
+                if src.startswith("/"):
+                    src = ROOT + src
+
                 if isinstance(content[-1], str):
                     content.append([src])
                 else:
@@ -39,9 +42,13 @@ def article(url):
                 content.append(t)
         else:
             print("Unknown :", type(ele), ele)
+    if confirm:
+        print(content)
+        if input("Download [y/n]? ").lower()[0] != "y":
+            return 
     with open(f"photo/data/zipai/{aid}.json", "w") as f:
         json.dump({"data": content}, f)
-    # print(aid, title)
+    print([str(aid), title])
     
 def worklist(uid):
     # url = f"{ROOT}/my/{uid}/"
@@ -72,14 +79,48 @@ def main(author):
 # article("/selfies/202105/121680.html")
 # article("/selfies/202105/121680.html")
 
-for uid in range(1000, 1100):
-    for title, href in worklist(str(uid)):
-        print(title, ROOT+href)
-    if input(f"Continue [y/n]?: ")[0].lower() == "y":
-        pass
-    else:
-        break
+# for uid in range(1000, 1100):
+#     for title, href in worklist(str(uid)):
+#         print(title, ROOT+href)
+#     if input(f"Continue [y/n]?: ")[0].lower() == "y":
+#         pass
+#     else:
+#         break
 
 
 # article("/selfies/201807/68971.html")
 # main("12765")
+
+
+while True:
+    inp = input("$ ").strip()
+    if inp.lower() == "exit":
+        print("Exit system!")
+        break
+    elif inp.lower() == "pwd":
+        print(f"pwd: {os.getcwd()}")
+    elif inp.lower() == "":
+        pass
+    elif inp.split()[0].strip().lower() in ["article", "author"]:
+        cmd, *args = inp.split()
+        if cmd.strip() == "article":
+            article(args[0], True)
+        # print(cmd, args)
+    else:
+        print(f"Command {inp} is not defined!")
+
+
+"""
+https://99zipai.com/selfies/202111/127295.html
+https://99zipai.com/selfies/202204/133807.html
+https://99zipai.com/my/13871/
+https://99zipai.com/my/13300/
+https://99zipai.com/my/14097/
+https://99zipai.com/my/4045/
+# https://99zipai.com/selfies/202202/130208.html
+https://99zipai.com/my/9926/
+https://99zipai.com/my/13411/
+
+
+
+"""
